pyproject.toml
# This Source Code Form is subject to the terms of the Mozilla Public
# License, v. 2.0. If a copy of the MPL was not distributed with this
# file, You can obtain one at https://mozilla.org/MPL/2.0/.

[tool.poetry]
name = "enviroplus-homeassistant"
version = "0.1.0"
description = ""
authors = ["Erwin de Haan <1627021+EraYaN@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "enviroplus_homeassistant"}]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"
enviroplus = "^0.0.5"
"RPi.GPIO" = "^0.7.0"
smbus = "^1.1.post2"
atmos = "^0.2.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
